LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/blob-data.https-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: blob:null/fc641426-a8aa-478d-8674-ec8d3ee287e1
CONSOLE MESSAGE: [blocked] The page at data:text/html,<!doctype html><script src=https://web-platform.test:9443/resources/testharness.js?pipe=header(Cross-Origin-Resource-Policy,cross-origin)%3E%3C/script%3E%3Cscript%3E/*%20Cannot%20use%20httpWorkerIncrementerTest()%20here%20as%20the%20HTTP%20URL%20is%20not%20same%20origin.%20*/promise_test(t%20=%3E%20{%20%20const%20worker%20=%20new%20Worker(URL.createObjectURL(new%20Blob([`const%20view%20=%20new%20Uint8Array(new%20SharedArrayBuffer(1));self.onmessage%20=%20()%20=%3E%20{%20%20const%20succeeded%...0%20});},%20%22data%20frame:%20postMessaging%20to%20a%20dedicated%20blob%20sub-worker%20allows%20them%20to%20see%20each%20others'%20modifications%22);test(()%20=%3E%20{%20%20assert_equals(self.origin,%20self.location.origin);},%20%22data%20frame:%20self.origin%22);test(()%20=%3E%20{%20%20assert_equals(self.crossOriginIsolated,%20true);},%20%22data%20frame:%20self.crossOriginIsolated%22);test(()%20=%3E%20{%20%20assert_true(self.isSecureContext);},%20%22data%20frame:%20self.isSecureContext%22);%3C/script%3E requested insecure content from blob:null/559aaf55-6d4c-4132-b6c2-5eb024a0e16a. This content was blocked and must be served over HTTPS.

CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Cannot load blob:null/559aaf55-6d4c-4132-b6c2-5eb024a0e16a due to access control checks.

Harness Error (TIMEOUT), message = null

PASS blob worker: postMessaging to a dedicated HTTP sub-worker allows them to see each others' modifications
PASS blob worker: postMessaging to a dedicated blob sub-worker allows them to see each others' modifications
PASS blob worker: self.origin
PASS blob worker: self.crossOriginIsolated
PASS blob worker: self.isSecureContext
TIMEOUT data worker: postMessaging to a dedicated blob sub-worker allows them to see each others' modifications Test timed out
PASS data worker: self.origin
FAIL data worker: self.crossOriginIsolated assert_equals: expected false but got true
PASS data worker: self.isSecureContext
TIMEOUT data frame: postMessaging to a dedicated blob sub-worker allows them to see each others' modifications Test timed out
PASS data frame: self.origin
PASS data frame: self.crossOriginIsolated
FAIL data frame: self.isSecureContext assert_true: expected true got false
PASS blob frame: postMessaging to a dedicated HTTP sub-worker allows them to see each others' modifications
PASS blob frame: postMessaging to a dedicated blob sub-worker allows them to see each others' modifications
PASS blob frame: self.origin
PASS blob frame: self.crossOriginIsolated
PASS blob frame: self.isSecureContext
